Move Mongo and Redis based config options to `virtool_workflow_runtime.config`.
